transferes announcement to dashboard; optimizes UI;
